# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EventHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{id}/register:
    post:
      operationId: registerForEvent
      tags: [registrations]
      summary: Register current user for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Already registered or event full
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /events/{id}/volunteers:
    get:
      operationId: listEventVolunteers
      tags: [volunteers]
      summary: List volunteer applications for an event (organizer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Volunteer applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VolunteerApplication"
  /events/{id}/volunteer-apply:
    post:
      operationId: applyToVolunteer
      tags: [volunteers]
      summary: Apply to volunteer for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolunteerApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerApplication"
        "400":
          description: Already applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}/tasks:
    get:
      operationId: listEventTasks
      tags: [tasks]
      summary: List tasks for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task for an event (organizer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Registrations ─────────────────────────────────────────────────────────
...
  /registrations/{id}:
    delete:
      operationId: cancelRegistration
      tags: [registrations]
      summary: Cancel a registration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cancelled
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Check-in ──────────────────────────────────────────────────────────────
...
  /volunteer-applications/{id}:
    put:
      operationId: updateVolunteerApplication
      tags: [volunteers]
      summary: Approve or reject a volunteer application (organizer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolunteerApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerApplication"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Tasks ─────────────────────────────────────────────────────────────────
  /tasks/{id}/assign:
    put:
      operationId: assignTask
      tags: [tasks]
      summary: Assign a volunteer to a task (organizer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskAssignInput"
      responses:
        "200":
          description: Task assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAssignment"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Registration:
      type: object
      required: [id, eventId, userId, status, qrToken, registeredAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [registered, waitlisted, cancelled]
        qrToken:
          type: string
        qrCodeDataUrl:
          type: ["string", "null"]
        checkedInAt:
          type: ["string", "null"]
          format: date-time
        checkedOutAt:
          type: ["string", "null"]
          format: date-time
        registeredAt:
          type: string
          format: date-time
...
    RegistrationInput:
      type: object
      properties:
        notes:
          type: string

    # ─── Check-in ─────────────────────────────────────────────────────────────
...
    VolunteerApplication:
      type: object
      required: [id, eventId, userId, status, appliedAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected]
        appliedAt:
          type: string
          format: date-time
        volunteerName:
          type: string
        volunteerEmail:
          type: string
        eventTitle:
          type: string
    VolunteerApplicationInput:
      type: object
      properties:
        message:
          type: string
    VolunteerApplicationUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]

    # ─── Tasks ────────────────────────────────────────────────────────────────
    Task:
      type: object
      required: [id, eventId, title, volunteersNeeded, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        stationLocation:
          type: ["string", "null"]
        startTime:
          type: ["string", "null"]
          format: date-time
        endTime:
          type: ["string", "null"]
          format: date-time
        volunteersNeeded:
          type: integer
        assignedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [title, volunteersNeeded]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        stationLocation:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        volunteersNeeded:
          type: integer
          minimum: 1
    TaskAssignment:
      type: object
      required: [id, taskId, volunteerId, status]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        volunteerId:
          type: integer
        status:
          type: string
          enum: [assigned, confirmed, completed]
    TaskAssignInput:
      type: object
      required: [volunteerId]
      properties:
        volunteerId:
          type: integer
...