# lib/api-zod/src/generated/types/taskAssignment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EventHub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskAssignmentStatus } from './taskAssignmentStatus';

export interface TaskAssignment {
  id: number;
  taskId: number;
  volunteerId: number;
  status: TaskAssignmentStatus;
}