# lib/api-zod/src/generated/types/volunteerApplicationUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EventHub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VolunteerApplicationUpdateStatus } from './volunteerApplicationUpdateStatus';

export interface VolunteerApplicationUpdate {
  status: VolunteerApplicationUpdateStatus;
}